backend/requirements.txt
fastapi>=0.110.0
pydantic>=2.6.3
uvicorn>=0.28.0
pytest>=8.1.1
pytest-httpx>=0.30.0 	# library for helping test async requests
pytest-asyncio>=0.23.5  # also for helping test async requests
httpx>=0.27.0
python-dotenv>=1.0.1
mypy>=1.9.0
pep8>=1.7.1